# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      account:
        description: 账号（邮箱或手机号）
        example: admin
        type: string
      password:
        description: 密码
        example: password123
        type: string
    required:
    - account
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        description: JWT token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        description: 用户信息
    type: object
  auth.RefreshTokenRequest:
    properties:
      token:
        description: 当前 token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - token
    type: object
  auth.RefreshTokenResponse:
    properties:
      token:
        description: 新 token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  common.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
  internal_handler_user.CreateUserRequest:
    properties:
      avatar:
        description: 头像
        example: http://example.com/avatar.jpg
        maxLength: 500
        type: string
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
      name:
        description: 姓名
        example: 张三
        type: string
      password:
        description: 密码
        example: password123
        minLength: 6
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        maxLength: 20
        type: string
      status:
        description: 状态：1-正常 2-禁用
        enum:
        - 1
        - 2
        example: 1
        type: integer
    required:
    - email
    - name
    - password
    type: object
  model.UserResponse:
    properties:
      avatar:
        example: http://example.com/avatar.jpg
        type: string
      created_at:
        example: "2026-01-08T10:00:00Z"
        type: string
      email:
        example: zhangsan@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_login:
        example: "2026-01-08T10:00:00Z"
        type: string
      name:
        example: 张三
        type: string
      phone:
        example: "13800138000"
        type: string
      status:
        example: 1
        type: integer
      updated_at:
        example: "2026-01-08T10:00:00Z"
        type: string
    type: object
  system.HealthResponse:
    properties:
      status:
        description: 状态
        example: ok
        type: string
      timestamp:
        description: 时间戳
        example: "2026-01-08T10:00:00Z"
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      avatar:
        description: 头像
        example: http://example.com/new-avatar.jpg
        type: string
      email:
        description: 邮箱
        example: lisi@example.com
        type: string
      name:
        description: 姓名
        example: 李四
        type: string
      phone:
        description: 手机号
        example: "13900139000"
        type: string
      status:
        description: 状态：1-正常 2-禁用
        example: 2
        type: integer
    type: object
host: localhost:1323
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 一个基于 Go 语言开发的 REST API 服务，使用 Echo 框架构建
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Gosir API
  version: "1.0"
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: 获取所有用户列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - 用户管理
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - 用户管理
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用账号和密码登录系统
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 将当前 token 加入黑名单，强制失效
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用当前 token 获取新的 token，延长登录时间
      parameters:
      - description: 刷新 token 信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.RefreshTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
      summary: 刷新 token
      tags:
      - 认证
  /health:
    get:
      consumes:
      - application/json
      description: 检查服务健康状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.HealthResponse'
      summary: 健康检查
      tags:
      - 系统
securityDefinitions:
  Bearer:
    description: 请输入 JWT token，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"